Add service to encapsulate API calls
